extract deposit confirm card into separate component
